backend/requirements/requirements.txt
django==5.1.1
Pillow==10.4.0
psycopg2-binary==2.9.9
requests==2.32.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-extensions==3.2.3
gunicorn==23.0.0
django-celery-results==2.5.1
celery==5.4.0
sentry-sdk[django]==2.14.0
redis==5.1.0
django_celery_beat==2.7.0
qrcode[pil]==8.0
python-telegram-bot==21.11.1